* master:
  //更改  ZHFService  请求方式 名
  //.购物车订单 本地化  加减处理
  //处理登录 本地储存 以及 登录逻辑处理 // 商品本地化 ，GRDB  删除所有 库 的处理

+++ Cashier/GRDB/ZWSementGRDB.swift
@@ -204,11 +204,18 @@ extension ZWSementGRDB:MutablePersistableRecord, FetchableRecord {
             }
         }
     }
-    
+    //MARK:删除所有
     static func deleteAll() ->Void {
-        // 是否有数据库表
+
         try! self.dbQueue.inDatabase { (db) -> Void in
-            try! ZWSementGRDB.deleteAll(db)
+            // 判断是否存在数据库
+            if try db.tableExists(TableName.ZWSementGRDB) {
+                debugPrint("表已经存在")
+                try! ZWSementGRDB.deleteAll(db)
+                return
+            }
         }
     }
+    
+    
 }
